Fix double/int bitmask in BuildWorld seed calculation
Time.GetUnixTimeFromSystem() returns double in the Godot 4 C# bindings,
so `(int)(Time.GetUnixTimeFromSystem() & 0x7fffffff)` fails to compile
with "Operator '&' cannot be applied to operands of type 'double' and
'int'". This was masked locally because nothing in this repo had ever
been a real C# build before -- the project just showed Godot's default
grey clear color.

Cast to long first so the bitmask is a long-vs-long operation, then
narrow to int.

+++ scripts/Game/Main.cs
@@ -47,7 +47,7 @@ public partial class Main : Node2D
 
 	private void BuildWorld()
 	{
-		int seed = (int)(Time.GetUnixTimeFromSystem() & 0x7fffffff);
+		int seed = (int)((long)Time.GetUnixTimeFromSystem() & 0x7fffffff);
 		GD.Print($"[world] seed = {seed}");
 
 		var world = WorldGenerator.Generate(seed);
